pyproject.toml
[tool.poetry]
name = "aou discord"
version = "0.1.0"
description = ""
authors = ["vivax <vivax3794@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
python-dotenv = "^0.19.0"
loguru = "^0.5.3"
aiohttp = "^3.7.4"
dnspython = "^2.1.0"

taskipy = "^1.8.1"
APScheduler = "^3.7.0"
motor = "^2.5.1"
hikari-tanjun = { git="https://github.com/FasterSpeeding/Tanjun" }

[tool.poetry.dev-dependencies]
flake8 = "*"
flake8-bugbear = "*"
pep8-naming = "*"
flake8-builtins = "*"
flake8-docstrings = "*"
darglint = "*"
flake8-isort = "*"
flake8-comprehensions = "*"
flake8-bandit = "*"
flake8-colors = "*"
flake8-eradicate = "*"
flake8-blind-except = "*"
flake8-black = "*"
flake8-broken-line = "*"
flake8-fixme = "*"
flake8-multiline-containers = "*"
flake8-pep3101 = "*"
flake8-simplify = "*"
flake8-variables-names = "*"
flake8-functions = "*"

flake8-use-fstring = "*"
flake8-length = "*"
dlint = "*"
flake8-annotations-complexity = "*"
flake8-type-checking = "*"

isort = "*"
black = "^21.7b0"
pyright = "^0.0.10"
pre-commit = "^2.13.0"
pymongo-stubs = "^0.1.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true

[tool.black]
line-length=80

[tool.taskipy.tasks]
bot = { cmd = "python -m bot", help = "Run the bot" }
lint = { cmd = "pre-commit run --all-files", help = "Lints project" }
precommit = { cmd = "pre-commit install", help = "Installs the pre-commit hook" }
format = { cmd = "isort .; black .", help = "Runs the black python formatter" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pyright]
include = ["bot"]
pythonPlatform = "linux"
pythonVersion = "3.10"

useLibraryCodeForTypes = true

strictListInference = true
strictDictionaryInference = true
strictSetInference = true
strictParameterNoneValue = true
reportFunctionMemberAccess = true
reportMissingImports = true
reportMissingModuleSource = "warning"
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportUntypedBaseClass = true
reportUntypedNamedTuple = true
reportPrivateUsage = true
reportConstantRedefinition = true
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportOverlappingOverload = true

# motor is not type hinted so it complains
# reportUnknownParameterType = true
# reportUnknownArgumentType = true
# reportUnknownLambdaType = true
# reportUnknownVariableType = true
# reportUnknownMemberType = true

reportMissingTypeArgument = true
reportInvalidTypeVarUse = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
